PortTranslator: Create MidoNet port for Neutron VIP ports
This allows a user to use non-MidoNet LBaaS providers like "haproxy".
Note: For "haproxy" provider, a VIP port is a real interface (veth)
which needs to be connected to our datapath as an exterior port.

This also fixes one of blockers for FIP association to VIP.

Note: We intentionally do not preseed ARP for VIP ports because it
would confuse the MidoNet-native LBaaS implementation.

Ref: IN-59
Ref: MI-500
Signed-off-by: YAMAMOTO Takashi <yamamoto@midokura.com>
Change-Id: Id9eef172f8ab124157a201c1dba2754114ea92a5