PortTranslator: Create MidoNet port for Neutron VIP ports
This allows a user to use non-MidoNet LBaaS providers like "haproxy".
Note: For "haproxy" provider, a VIP port is a real interface (veth)
which needs to be connected to our datapath as an exterior port.

This also fixes one of blockers for FIP association to VIP.

Note: We intentionally do not preseed ARP for VIP ports because it
would confuse the MidoNet-native LBaaS implementation.

Ref: IN-59
Ref: MI-500
Signed-off-by: YAMAMOTO Takashi <yamamoto@midokura.com>

Conflicts:
	docs/openstack-lbaasv1.md
	midonet-cluster/src/main/scala/org/midonet/cluster/services/c3po/translators/PortTranslator.scala
	midonet-cluster/src/test/scala/org/midonet/cluster/services/c3po/translators/PortTranslatorTest.scala
	specs/2015.03/neutron_translation.md

Change-Id: I3e9553c92d14b68098b3431b85772e3d096a43a5