Shutdown VT executors when stopping the agent
This is a minor improvement where the VT executors are shutdown
after stopping the agent services. This ensures that after this
point the VT stops processing updates from NSDB. The executors
are shutdown after the agent services, to ensures that any service
that requires VT updates for a graceful shutdown, receives them.

The patch also updates the VT executor check function, to validate
the actual VT thread.

Change-Id: Ia344258df35b86ff8e41a73468706b6595092d06
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>