MDTS: skip non-necessary tests in basic neutron suite
The basic neutron test suite was meant to be used as a documentation on
how to use the new mdts setup/teardown mechanism. Disable most of the
tests (keep one for sanity check) and fix a small bug by which the mac
address is not the same in the binding and the virtual topology
definition (making the ping to fail in nightly jobs).

Ref.: MI-612
Change-Id: I392f3e7cc913609571da925415396c35c318bac7
Signed-off-by: Xavi León <xavi.leon@midokura.com>