Update patch set 1
Patch Set 1: Code-Review+1

When I mean neutron developers I mean our plugin's developers. I agree that we should use the released version of midonet for the sake of stability, but the reality is that plugin developers have to tip midonet and neutron master branches together. Don't know the reasons. 

I'll set to +1 because it is an improvement to current  approach.

Patch-set: 1
Label: Code-Review=+1