RHEL: missing java7 path from search dirs
Midolman didn't start when a java version different from java7 was set
as default (e.g. default set to java8). This was because the actual
java7 patch was missing from the search directories, so it would never
find it and failing to start. This patch adds the missing directory
to the search paths.

Ref.: MI-627
Change-Id: Ic7c4fc2976836c907dc2f5d57c91cd111b009766
Signed-off-by: Xavi León <xavi.leon@midokura.com>