Reply DHCP nack upon incorrect DHCP request
Reply with a DHCP NACK when an incorrect DHCP REQUEST is sent. This
might happen when a client holds an outdated lease and tries to renew
it. Up until now, the client had to wait for the request to timeout
(usually 60s). This patch sends a NACK (as defined in the RFC-2131)
upon an incorrect request so the client can issue a DHCP DISCOVER right
away without having to wait. Added an MDTS test that covers this
use case.

Ref.: MI-536
Signed-off-by: Xavi León <xavi.leon@midokura.com>

Change-Id: Ib0ad8304a484162f2c77055ade161c55d9d1f398