Relax SG rules for IP fragmentation (L4 rules)
For SG rules which need to examine L4 fields,
allow a fragmented packet if:

- it's the first fragment (offset == 0) and it's big enough to include
  the complete L4 header, and the header is matching the rule,
- or it's a non-first fragment (offset > 0) and it's matching the rule,
  ignoring L4 part of the rule.

The second case might be controversial, as it allows packets which
might not actually match with the original SG rule.  Note that these
fragments won't be reassembled into a complete datagram, unless
the first fragment is passed as well.  This is a loose-but-efficient
approach.  A more strict alternative would involve IP defragmentation
before applying rules, which might or might not be a good idea.

Ref: MI-573
Signed-off-by: YAMAMOTO Takashi <yamamoto@midokura.com>
Change-Id: I479b7d045651777991bb66671b1eb1ff765fc7dd
(cherry picked from commit 74d63b43a47e54e30d3acbc994af3b4d42749153)