VPNaaS: Introduce an executor pool
This introduces a new executor pool for the containers service in
the agent, which will allow different containers to use distinct
single threaded executors. The number of threads in the pool is
configurable, or it defaults to the number of available processors
if the configuration is not customized.

The pool allows a caller to fetch one executor from the pool at a
time, ensuring that consecutive calls return different executors
from the pool.

Ref: MI-665

Change-Id: I45e62fb3466a4b405fa3789f861293a6895f369b
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>