VPNaaS: Replace Scala process with custom implementation
When using the Scala process builder, for every command we execute
from the IPSec container handler, the builder will create 3 new
threads, with only purpose of managing the input/output streams
exposed by the new process.

For the IPSec container, this is an unnecessary overhead, because
of the following:
- execution of the external process is synchronous with the
container handler thread (i.e. the contaier handler thread is
blocked until the child process exits)
- we only use the stdout and stderr input streams, and both streams
are used in the same way

Therefore, this patch replaces the scala process builder with
a custom implementation that:
- redirects the stderr stream to stdout such that we only need to
read a single stream
- implements a local loop that reads the data from the output
stream on the container handler thread while the child process is
executing

Change-Id: I651039f00301a1788f2fb0be9f579d4b1fef664e
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>