VPNaaS: Identify container handlers
Currently there is no mechanism to distinguish between container
handlers that correspond to the same physical container. With this
change the container provider allows to create a container handler
instance that is injected with an UUID argument annotated with a
Guice names annotation, where the name is `id`.

The same new mechanism is used to provide an executor service per
container handler, allowing different containers to use different
singe-threader executors.

This will be used in a subsequent patch to ensure that container
instances are easily identifiable with respect to whether they
correspond to the same physical container.

Ref: MI-665

Change-Id: I25a39a09e894fc4bc447fdfeb2e0498bba74a6be
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>