VPNaaS: Container logger uses container identifier
This modifies the container logger as follows. Instead of using
an empty file, where the file name is <name>.<type>, the logger
now uses a file name <id>.<type> where the container name is
written to the file.

This helps identify the log file as corresponding to a particular
container.

Ref: MI-665

Change-Id: If17db94252ce7c752c9876469954db121100571b
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>