VPNaaS: Containters are multi-threaded
This change replaces the single conveyor belt from the container
service with the container context. There is one container context
per container identifier, such that all handlers created for the
same container identifier (i.e. the same container port) will
share the same container context.

The container context gives the single-threaded executor on which
the container executes, while the container service reserves a
single-threaded executor for itself to receives notifications
from NSDB and manage the containers.

The container context ensures that all operations executed on a
container are serialized, while the executor remains avaialble to
complete asynchronous container tasks, or to be used in parallel
by other containers.

Notifications from the containers are executed on the container
context thread, and where needed (e.g. logging) they are
synchronized.

Ref: MI-665

Change-Id: If182af4d5b0786f976ee2339feb8787c60caa288
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>