Update patch set 8
Patch Set 8:

I had time to think through this a bit more and now I'm thinking, if we remove the restriction that we only apply FIP DNAT/SNAT on traffic ingressing/egressing the gateway port, our currently implemented rules make all cases work.  The reason why such restriction exists, IIRC, is that we had a false notion that NAT should not be applied for E-W traffic, but that doesn't match the reference implementation, and influencing implementing other FIP scenarios.  We already have an SNAT rule added on router-interface-add event for traffic ingressing/egressing the same port, which was needed for the same purpose as non-FIP to FIP case, but with VIP.  Thoughts?

Patch-set: 8