VPNaaS: Introduce a container context
A container context provides a single execution queue for the
operations to execute inside a container. The context receives a
single threaded executor, that is responsible for executing the
container tasks, where some tasks may complete asynchronously,
on the context thread or on a different thread.

The context will execute a new task (either newly submitted or
waiting in the queue) only when the previous task has completed
its future. This makes the context executor available for other
tasks, such as a different container context or even to complete
the pending task.

Ref: MI-665

Change-Id: I9f8009c33cee86c8cebf22b0565f4ee824e7da3d
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>