Update patch set 8
Patch Set 8:

after discussing with fernando, we came up with one scenario that does not work, which is priv IP to priv IP connection when FIPs are associated.  In that case, SNAT needs to be avoided.  To do so, in the reference implementation (iptbles) it has a rule that says, 'if ! DNAT ACCEPT'.  Basically it tracks whether DNAT has happened in the forward flow and avoids SNAT unless DNATed, and we do not have this capability.  My question now is whether we should simply implement this instead of finding a way around it.

Patch-set: 8