Update patch set 1
Patch Set 1:

> (3 comments)

So, this is a backport cherry-pick for an already existing fix, so I am wondering if these comments should have been applied to the original fix, or are these something special to this branch in particular?  If the former, I don't think it's appropriate to change these here and then re-forward-port the back-port, rather we should revert and amend the main fix and re-backport those again.  what do you think?

Patch-set: 1