MI-522: python-midonetclient support for HATEOAS
Since only the support of absolute path URL is needed, modification
of auth_lib.Auth seems fair enough to not make a big impact, despite
it seems to violate the single responsability principle, and it would
fit better in a class/code not related with authorization function.

Finally, despite relative URL are not going to be supported,
URLs relative to base URI are accepted (i.e: URLs relative to the
service or web application context) as a fallback non-standard option.

Auth class constructor changed to accept the base URL and the app-path
('login' in this case), so that the base URI can be passed without
ambiguity.

_resolve_uri method is used to resolve the URL passed by callers
in the do_request and do_upload methods.

Change-Id: I608287b7a5eb9f856804a15ca969a48eec55f767
Signed-off-by: Miguel Ángel Herranz Trillo <miguel@midokura.com>