FlowState: introduce cluster minion
This patch introduces the cluter minion that receives and exposes
flow state messages from and to MidoNet Agents. This minion registers
itself as an instance of the flowstate service so agents can get its
connection information to send and retrieve flow state. This
initial patch does not consider how it will store those messages -- it
could be in cassandra or memcached and thus this service could be
executed on any node, or in-memory in a single cluster node and thus a
leader election mechanism should be put in place.

Change-Id: I851a4f3234b351b49d898b72f3945f0b44e7bfeb
Signed-off-by: Xavier León <xavi.leon@midokura.com>