Update patch set 7
Patch Set 7:

(1 comment)

Seems to me that using observables would make this more readable.

I'm also concerned about the potential load this would put on zookeeper. I've seen components like this negatively impact performance of systems in the past due to trying to clean up too fast.

I don't see anything to stop multiple cluster nodes running the recycler at the same time. That doesn't need to be in this patch though (but this patch should disable by default if it's not there).

Finally, this is more of a Reaper than a recycler.

Patch-set: 7