Update patch set 7
Patch Set 7:

(1 comment)

- Observable more readable

Correct, we already have a method called asObservable wrapping the curator BackgroundCallback with an observable. It it okay this overhead?

- Load on ZK

Should we throttle the cleanup operations? If so, what should be the default tx rate?

- Synchronization between different cluster nodes.

The current code already does this by leveraging the root zoom znode, to determine whether the cleanup can start. Only one node can succeed in modifying the node while still matching the start time. This doesn't work if nodes use different clocks, or if a recycling operation takes too long and it overlaps with the next one (either from the same node or a different node) - so the guarantee of a single cleanup at a time is weak and it depends on configuration. Do we want a stronger check, and therefore use leader election?

Patch-set: 7