Agent: Be resilient against duplicate linked flows
This patch ensures that a pair of linked flows are indexed
independently, so they are assigned different netlink sequence numbers
when created in the datapath. Linked flows now reference each other,
so that if a flow is detected as a duplicate we can remove the linked
flow from the datapath.

Ref MI-625

Change-Id: Ifbd91dabebb43728c044e60f7d1c23a8271d54be
Signed-off-by: Duarte Nunes <duarte@midokura.com>
(cherry picked from commit 22fbfa99367dd77a3c40eb875495d85f5bf2f5aa)