Agent: Don't rely on concurrent queue size method
The ArpRequestBroker unit test failed with a Match error with the
processNewMacs() protected by a call to macsDiscovered.isEmpty(), that is,
macsDiscovered.poll() returned null even though isEmpty() was false.
Since size() implementations in concurrent queues are usually weak,
the return value of poll() can never be assumed to be consistent with
size(). So, this patch changes the implementation of processNewMacs()
to not rely on isEmpty() and thus on size().

Change-Id: Ibda2c53008feb9ff72d69b20ae8dbd012606ac8b
Signed-off-by: Duarte Nunes <duarte@midokura.com>