Operator hint in failed to create host log message
When renaming a host while keeping the same UUID, the agent fails
to start because the existing object from the topology does not
match the new name. The possible workarounds are:

* Delete the host UUID file: this would create a new host object,
  while keeping the previous one.

* Delete the host from the topology: this would allow the agent to
  create a new entry with the new name. The operator will also have
  to re-add the host to the tunnel zone, and re-set any host settings
  such as the flooding proxy and container weights.

This adds the second solution to the log message, to give the
operator a quick hint wrt the solution, as well as refer to the
documentation for more information.

Change-Id: I3b51787da9e3286c06e70cef829b6e5ef880527d
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>