Mock directory: watchers trigger according to type
This fixes an issue in the mock directory, where all watchers for
a node would trigger, regardless of the event that modified the
node. With this change the node watchers are divided into three
buckets, and they are triggered as follows, consistent with the
behavior of the ZooKeeper client:

* Exists watchers trigger on node create, delete and set data

* Get data watchers trigger on node delete and set data

* Get children watchers trigger on node delete, add child and
  delete child.

This change is necessary to accommodate further unit tests for the
detection of object deletions in the cluster managers.

Ref: MI-561

Change-Id: Icd8ac70bf4d069479e6195ae28883ddf49e288c7
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>