Update patch set 1
Patch Set 1: Code-Review-1

(9 comments)

Could you update neutron_translation.md?

Generally LGTM but I found a lot of unrelated code changes in this patch that made it slightly more difficult to review.  Ultimately, I won't -1 for that, but if you could separate the 'clean-up' patch from the actual change for this fix, it will be better.

Patch-set: 1
Label: Code-Review=-1