Fix NPE in L3ZkManager.prepareDeleteGatewayPort
We've encountered a situtation where Neutron deletes a tenant router before
deleting its gateway port and then tries to delete the router gateway port. This
results in an NPE when we try to fetch the peer of the port on the provider
router and it doesn't exist. We don't know why this happens, but this change
guards against the NPE by checking for null and returning so that the port on
the provider router can be safely deleted.

Change-Id: I5e1cbf9d4c04ac295b49edfc2e9bf8a33d60b48f
Signed-off-by: Brandon Berg <bberg@midokura.com>
Fixes: MI-709