Agent: Don't use Stash for recircMatch
We were using an InstanceStash to fetch the recircMatch when
encapsulating or decapsulating a packet. However, recircMatch is used
by the FlowProcessor during the output stage, while the lifetime of a stashed
instance is limited to the simulation stage. This patch fixes this by
lazily initializing and caching the recircMatch in the PacketContext.

Change-Id: Ibe898ac7bd4b9becedbae2fba75df41e9712ad3c
Signed-off-by: Duarte Nunes <duarte@midokura.com>