External subnet update checks gateway is set
When updating a Neutron external subnet, we use the previous
gateway IP address to search for the peer gateway router port that
is configured with that IP such that we update the IP address of
that port as well if it has changed.

However, the update method does not check if the previous gateway
IP address is null, which eventually throws an exception.

This patch addresses this missing validation, by allowing the
creation of a subnet without a gateway IP address, and subsequently
updating it with a valid address. This change also ensures that
the external network is linked to the provider router only when
it is configured with a gateway IP address. If during a create or
update, the gateway IP is removed, then the Neutron API unlinks
the subnet from the provider router.

Ref: MI-730

Change-Id: Ic0301cfd6fa4c4153c0442b2a7def52a7a80da3d
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>