Field forwardChainId in Router DTO missing converter
In MidoNet ZoomConvert automatically converts between Java and
Protocol Buffers UUID fields. However, this improvement is not
available in MidoNet 5.0, where classes converter to/from Protocol
Buffer messages must specify the converter for UUID fields.

A recent change in the Router API DTO has added a new UUID,
probably as part of a backport. However, it didn't add the
required converter and therefore conversions of this class will
fail with ZoomConvert.

This patch addresses this error.

Fixes: MI-733

Change-Id: I7f699301a3f498732e0fce9d3de6ce4b62f1f5ce
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>