CLI: ensure application clears app on exception
Before each API request the MidoNet CLI executes a methods that
downloads the set of base URIs from the API. All subsequent
requests for resources are made using this list of base URIs. This
list is delivered as an Application JSON object, which is
initialized to an empty object, and then populated with the fields
received in the API response.

However, if there is an exception when requesting the Application
DTO, the CLI handles the exception and resets the Application
object as undefined, but only if the exception was a socket
`connection refused` error.

The CLI determines a socket error is connection refused by
comparing the error message with `ECONNREFUSED`, which for recent
versions of Python is something like:

   `error: Errno 111 Connection refused`

In any case there are other exceptions too with various error codes
corresponding to a failed connection (e.g. destination unreachable,
connection reset by peer, etc) for which the CLI should use the
errno codes rather than the socket error message.

This patch does the following:

* Removes the MidoApiConnectionRefused class that does not serve
any other purpose, and is incorrectly matched to the socket error
to begin with.

* Ensures that the Application object is reset to None after any
exception that occurs when requesting it to the API. This
guarantees that subsequent requests will succeed in getting a
fresh copy of the Application DTO if the API is subsequently
available.

Fixes: MI-638

Change-Id: Ic4ba7ee0c555f0345ecf52733e846d898f132931
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>
(cherry picked from commit f88fa640dc4cb66418352fe5d52884019836609e)