Make portsDelete of LocalDataClientImpl robust
portsDelete method fails if there are leftover data from one of the
previous versions of MidoNet.  In particular, when

/filters/<portId>/snat_blocks

is present, portsDelete fails with NodeNotEmptyStateException.

In order to make portsDelete method more robust so that users can clean
up stale data, use ZkOpList that already handles recursive deletion of
child nodes when NodeNotEmptyStateException is encountered from a
'multi' call for port deletion.

In addition remove 'delete' method from PortZkManager because it is no
longer used after this change.

Change-Id: Ic334fbee3836cfa4cc67a3c6a59f1d193ff778ba
Fixes: MI-352