Ensure conversion for all fields for rule conditions
This fixes an issue in the conversion of rule DTOs to/from
Protocol Buffers messages in the REST API, where certain fields
(VLAN fields) were not being converted. For this purpose, the
patch leverages the ability of Zoom convert to specify a target
class for which the conversion is performed and to eliminate the
need to copy manually the fields from super classes.

In addition, the patch strengthens the Rule API unit tests to
ensure we vaidate that a created rule is the same as after being
requested from the REST API. The patch also modifies the unit test
for validating the rule conversion at the agent side to consider
the missing fields.

Fixes: MI-699

Change-Id: I060fe4a6392292ac28fc82f1c26231941e44a5dd
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>