Update patch set 1
Patch Set 1:

Ivan: the difference here is that the lock previously obtained from the lock factory returns a boolean that says whether the lock was acquired successfully or not. We were not checking the returned value. With this patch we call acquire on a ZkOpLock which throws an exception if the lock is not acquired (and we perform the acquire in a try block).

Patch-set: 1
Reviewer: Nicolas Schiper <1002820@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0