Match on port when matching on ipaddr group
In order to implement security group rule that checks against another
security group where ovelapping IP addresses may exist, matching on
only the IP address is not enough.  This patch enhances the IP address
matching logic to also include the ingress and egress port check.  More
precisely, when matching on the IP address group source, also check the
ingress port ID.  Likewise, when matching on the IP address destination,
check the egress port ID.

There is still one situation that is still not addressed, which is when
you add a rule to match on the IP address group destination, depending
on where you add the rule, egress port ID might not be known.  In that
case, only IP address will be used to reserve the previous behavior.
This case will be handled in a separate patch.

Change-Id: I5b6d4819812fd3370c3c6ce6d2279ce320152b5d
Fixes: MI-377