InterfaceScanner: Handling IPv6 link-layer addresses
This patch handles interfaces that have an IPv6 link-layer address. In
this case the MAC contained in the Link object is null (rightfully
so). In this case, we set the MAC address in the InterfaceDescription
to null as well. As a result, the correspondign interface will be
filtered out, which is ok in this case.

Ref: MI-732

Change-Id: Ib818195960008efff24e7cc564a670d8ba4892d0
Signed-off-by: Nicolas Schiper <nicolas@midokura.com>