InterfaceScanner: Handling null MAC addresses
This patch handles interfaces that have L2 addresses that are not
6-bytes long. In this case the MAC contained in the Link object
is null (rightfully so). We thus set the MAC address in the
InterfaceDescription to null as well. As a result, the
correspondign interface will be filtered out, which is ok in
this case.

Ref: MI-732

Change-Id: Ib818195960008efff24e7cc564a670d8ba4892d0
Signed-off-by: Nicolas Schiper <nicolas@midokura.com>
(cherry picked from commit 2c890c14bced8ab968a464b77a5f88da87fa45af)