VXGW: Disconnect synchronizer when VTEP deleted
This fixes an issue in the VXLAN Gateway Service, where the VTEP
synchronizer would not close the VTEP connection when deleting the
VTEP from the NSDB topology. This is because under certain race
conditions the synchronizer was unsubscribed from VTEP notifications
before receiving the onCompleted notification from the VTEP
observable.

This patch ensures that we call the onCompleted method of the
synchronizer before it is unsubscribed from the update stream.

Ref: MI-717

Change-Id: I366f8cf890f555d66838acddb70dca8c687b6e0e
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>
(cherry picked from commit b36cb3044fc3102d175b4655089d584d32d850e2)