Update patch set 8
Patch Set 8:

(1 comment)

I've removed those implicits and replaced by explicitly calling apply. Also, FlowStatePackets is a typed trait now and the FromSbe methods (deserialize) is different on nsdb and the agent modules (similar to a factory as you suggested). ToSbe is not necessary as both extend *Store objects. 

Only left the implicits for the value classes that does not allocate.

The only extra allocations are done when fetching the objects from cassandra (which could be removed following a similar approach on a later patch)

Patch-set: 8
Label: Code-Review=0