Reverse dnat transformations in ICMP data
If after applying dnat to a packet, the packet cannot then be routed to
the correct destination, we should send an error back to the originator
of the packet, with the original packet as the icmp data. However, as
dnat has been applied to the packet, the sender of the icmp response is
not able to set the icmp data correctly.

This change fixes this at multiple points.

Firstly, a new rule is added to the FIP translations so that if the
destination address inside the icmp data is the fixed address from the
FIP, the icmp data will be modified to have the correct
address (i.e. the FIP).

Secondly, the icmp data, when generated by a router is generated from
the packet as it appeared when it entered the router, rather then how it
looked after traversing prerouting chains, as the prerouting chain can
modify the destination. This is needed because when a router sends a
icmp error it emits it directly to the port without going through its
outchain, so no compensatory rules could be applied.

Ref MI-187

Change-Id: I20b8c239309f69bb25d301f00459c61fd4756870
Signed-off-by: Ivan Kelly <ivan@midokura.com>