VXGW: Serialize MAC updates
This fixes an issue in the VXGW MAC updater, where the requests to
the VTEP OVSDB are not serialized. The change from this patch
ensures that an OVSDB operation executed from the MAC updater
completes before requesting the next one.

This patch may not eliminate completely races caused by stale cached
OVSDB tables. For such conflicts we may retry an OVSDB operation
that failed.

Fixes (partially): MI-590

Change-Id: Ibe7fd001998658e6a6c812ee98fcb78d76bba0cd
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>