Set pool member status in GET requests
Neutron pool member has a status field that was never set in the GET
requests.  Without this field set, Neutron LBaaS API cannot get member
status information to the users.  Neutron LBaaS driver for MidoNet will
do a GET of mebmers with a list of UUIDs of pool members.  In the
returned list of Member objects, their status field should be set to
the status of corresponding MidoNet PoolMember object.

This patch includes the following changes:

 * MemberResource accepts query string parameter, 'id', representing
   the member IDs to fetch.
 * NeutronZoomPlugin implements list member method that accepts a list
   of member ids to fetch.
 * NeutronZoomPlugin sets the member status of the fetched members.

With this change, you can send the following request:

 * GET /members?id=<memberId1>&id=<memberId2>& ...

This API is intended to be called from Neutron.

Change-Id: I7337fc395d24159f0903dfcb30d81d037ff7997d
Fixes: MI-389