Add exists check to Translator.translateDelete()
Neutron sometimes gets out of sync with Midonet and tries to delete things that
have already been deleted. This change adds a call to storage.get() in
Translator.translateDelete() to get the Neutron object. If it fails with a
NotFoundException, translation is skipped and the operation returns. If the get
succeeds, the Neutron object is passed to the specific implementation, e.g.
NetworkTranslator.translateDelete().

Change-Id: I5c9c4cb7636d97cf914f65e611e6867cdb0966c8
Fixes: MI-759
Signed-off-by: Brandon Berg <bberg@midokura.com>