remove '-sf' flag from haproxy command
This flag is dangerous because it kills the process whose pid
matches the contents of the pid file associated with a given
haproxy instance. This can happen when the node hosting an
haproxy instance goes down, then later it serves the same haproxy
instance. When the haproxy instance comes up the second time, it
will find the pid file with a pid that might point to some other
random process.

Fixes MI362

Change-Id: I476a53a56e5700edab0930f53721011b0a947459