Update patch set 2
Patch Set 2:

Hi Xavi, thanks for reviewing.

I thought it would be best to avoid deleting files from the file system because the location of the l4lb files are configurable by the user. If the user puts them in, say, /etc/midolman, then we can't blindly delete the whole directory. We would have to find pid files that look right. Not impossible at all, but I thought it was adding more complexity than it is worth to a sub system that is already overly complex and difficult to test. Please let me know if your opinion does not match mine.

I uploaded a new patch that proposes an alternative to removing the files: when updating a config, call killHaproxyIfRunning, which will send a term/kill signal to the current haproxy process. This should have the same affect as the '-st' flag, without the risks.

There is another risk to keeping the '-st' flag, but I don't know how likely it is. If haproxy dies for whatever reason, then the pid in the pid file isn't 'clean' anymore so we can't trust it. I don't know why this would happen, but if it did we are in trouble because the '-st' flag would cause haproxy to kill whatever pid is in that file.

Let me know what you think.

Patch-set: 2
Label: Code-Review=0