Do not use AKKA actor `ask` pattern in the VTA
In the current implementation, the VTA keeps track of in-progress
requests using an internal map, where the the actors sending
requests to the VTA are added. When a requests for a device completes
because the device has been received, the VTA uses this map to
send a message to each requesting actor waiting for that device.

This approach has a disadvantage. If a request for a device never
completes, because the device has been deleted or because it
simply does not exist, the corresponding sender actors are kept
forever in the sender map, in anticipation that the requests will
eventually be fulfilled.

Currently, the VTA has no mechanism of being notified of non-existing
or deleted devices, and this capability will be added later in
subsequent patchers.

The actors from the senders map are added from two different sources,
which use the same DeviceRequest message to the VTA: (i) actors
that send the request directly using `tell`, and (ii) the
underlying PromiseActorRef used by the `ask` mechanism, called by
`tryAsk`. This second type of hidden actors are responsible for
completing the futures returned by `tryAsk`. If the requested
device does not exist, the futures will timeout, however the
actors and all internal objects they reference will be stored
"forever" in the senders map.

This patch changes the `tryAsk` mechanism by avoid using the AKKA
`ask` pattern that creates the underlying PromiseActorRef. Instead
the `tryAsk` will create a new promise and send it along with the
request to the VTA. The VTA will store these promises in a separate
map, and will handle internally the future expiration, such that
when a future expires, the corresponding promise is removed and
its memory can be reclaimed by the GC.

Ref: MI-561
Ref: MI-672

Change-Id: I4f703e96bc52a8701dadc54057dd32b4cd3463d3
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>