VPNaaS: Introduce host container limit
This patch introduces a container limit per host in the topology
model. The container limit configuration uses the following
fields.

* A `container_limit` indicates the maximum number of containers
that a host can start. A negative value indicates that there is
no limit for the number of containers. A zero value indicates that
the host cannot start any containers (this is the same as setting
the `container_weight` field to zero). Any positive value
indicates the maximum number of containers that a host can start.

* An `enforce_container_limit` field that indicates whether the
agent will stop containers when decreasing the container limit.
When false, an agent will run more containers that the container
limit, if those containers were started when the limit was
greater. When true, an agent will ask the cluster to migrate
a number of containers to a different agent, and the cluster shall
attempt to do so, but only the containers that do not have host
affinity can be migrated.

Ref: MI-771

Change-Id: I6a797a04d1a521c040b9922bc9cd6cadb5d82c1c
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>