Update patch set 4
Patch Set 4: Code-Review-1

(15 comments)

Just some comments on README.  I think for this type of work, README is VERY useful, so I want to make sure we have something very clear.  We will eventually probably have to move it to the operational guide though, so I'm ok if we do it separately.

Patch-set: 4
Label: Code-Review=-1