Update patch set 3
Patch Set 3:

(1 comment)

iirc, if haproxy dies for whatever reason, then the -st flag doesn't matter at all. When the new haproxy instance comes up, it will override the pid file with the new pid (not sure about this, but that's what I remember when I looked at it).

Patch-set: 3