Introduce alternative port mirroring positions (nsdb and simulation)
This is planned to be used for Tap as a service (TaaS), for which
the reference implementation taps packets at the network side
of Security Groups.  (vs VM side)
In MidoNet, a Neutron network and Security Groups translate to
a bridge and bridge ports' in/out filters.  and thus the current
mirroring position is not appropriate.

The new set of mirroring positions are introduced instead of
moving the old ones, because the old ones will likely be used
by the future versions of TaaS as well.

Also, while I'm here, make Port's continueIn/continueOut private.

Ref: MNA-911
Ref: MI-703

Signed-off-by: YAMAMOTO Takashi <yamamoto@midokura.com>
Change-Id: Ia125a48bd76a87247c3f9e9d0a1a96b3b44a9306