VPNaaS: Introduce container service status quota
This adds a new field to the ContainerServiceStatus message,
called `quota`. This indicates the number of containers that a
host may start, which a host reports according to the number of
containers currently running at that host and the container limit.

If the quota is negative, there is no limit to the number of
containers a host can run. A quota of zero, indicates the host
has reached its container limit, and is not eligible of starting
additional containers.

Ref: MI-771

Change-Id: Id26f2d4f96edaf2022f974eeaaf4e1ee90fba95d
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>