VPNaaS: Container service reports container quota
With this change, the agent container service exposes both the
host container weight and the container quota via the container
service status to the cluster. To this end, the service monitors
all relevant container information from the host object, and
updates the service status, whenever the host changes and when
starting or stopping a container. The quota will reflect the
container limit without the number of running containers, unless
the limit is negative (no limit), in which case the quota is
negative too.

Ref: MI-771

Change-Id: If7554eac46ea5e5db0d237d574986b83151f91ec
Signed-off-by: Alex Bikfalvi <alex.bikfalvi@midokura.com>