API: reduce ZK lock contention (sync access to ZK lock per api node)
When doing a request to the neutron api, we acquire a ZK lock to
maintain consistency in case more than one component is writting to the
NSDB. This lock has a hard coded timeout of 5s.

When a non-trivial amount of concurrent requests are made, we have
observed contention to acquire the ZK lock and thus, those requests that
take more than 5s to get the lock fail.

This patch synchronizes the access to the lock by providing two
different implementations of the neutron api: a lockless and a locked
version. The locked version just calls the lockless version on the
methods that needs a ZK lock and its access is synchronized so different
requests (threads) will wait for the lock to be released before trying
to acquire it.

NOTE: we do not hold the lock for more than one operation because
requests could be load balanced among several api nodes.

Ref.: MI-404
Change-Id: I416f3e1eec3e52df3d32cf450f9d04cbfbe23a3d
Signed-off-by: Xavi León <xavi.leon@midokura.com>